refactor: move deck import to dedicated feature slice
