+++ src/pages/deck-import/ui/DeckImportView.spec.tsx
@@ -10,7 +10,7 @@ import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom/vitest";
 import { describe, expect, it, vi } from "vitest";
 
-import type { DeckImportPreview, DeckImportResult } from "@/features/deck/import";
+import type { DeckImportPreview, DeckImportResult } from "@/features/deck-import";
 
 import { DeckImportView } from "./DeckImportView";
